Version 1.1
1:fix the problem about only one instance where invoke a method
2:add multiply response type, example xml,image,html
3:change the json serializater from newtomsoft.json to litjson

+++ demo/SolutionAJaxTest/TestBLL/User.cs
@@ -23,7 +23,7 @@ namespace TestBLL
         /// <summary>
         /// 年龄
         /// </summary>
-        public int Age { get; set; }
+        public System.Decimal Age { get; set; }
 
         /// <summary>
         /// 生日
